+++ CustomizeAnywhere/Framework/ModConfig.cs
@@ -4,22 +4,11 @@ namespace CustomizeAnywhere.Framework;
 
 class ModConfig
 {
-    public bool canAccessMenusAnywhere { get; set; }
-    public SButton ActivateButton { get; set; }
-    public SButton customizeButton { get; set; }
-    public SButton dyeButton { get; set; }
-    public SButton tailoringButton { get; set; }
-    public SButton dresserButton { get; set; }
-    public bool canTailorWithoutEvent { get; set; }
-
-    public ModConfig()
-    {
-        this.canAccessMenusAnywhere = true;
-        this.ActivateButton = SButton.LeftShift;
-        this.customizeButton = SButton.D1;
-        this.dyeButton = SButton.D2;
-        this.tailoringButton = SButton.D3;
-        this.dresserButton = SButton.D4;
-        this.canTailorWithoutEvent = false;
-    }
+    public bool CanAccessMenusAnywhere { get; set; } = true;
+    public SButton ActivateButton { get; set; } = SButton.LeftShift;
+    public SButton CustomizeButton { get; set; } = SButton.D1;
+    public SButton DyeButton { get; set; } = SButton.D2;
+    public SButton TailoringButton { get; set; } = SButton.D3;
+    public SButton DresserButton { get; set; } = SButton.D4;
+    public bool CanTailorWithoutEvent { get; set; } = false;
 }