+++ HayBalesAsSilos/Framework/ModConfig.cs
@@ -4,5 +4,5 @@ internal class ModConfig
 {
     public bool RequiresConstructedSilo { get; set; } = true;
     public int HayPerBale { get; set; } = 240;
-    public int HaybalePrice { get; set; } = 5000;
+    public int HayBalePrice { get; set; } = 5000;
 }