+++ MultiplayerModChecker/Framework/ModConfig.cs
@@ -2,6 +2,6 @@ namespace MultiplayerModChecker.Framework;
 
 internal class ModConfig
 {
-    public string[] IgnoredMods { get; set; } = { "Cherry.MultiplayerModChecker" };
+    public string[] IgnoredMods { get; set; } = ["Cherry.MultiplayerModChecker"];
     public bool HideReportInTrace { get; set; } = false;
 }