+++ TrainStation/Framework/ModConfig.cs
@@ -2,8 +2,8 @@ namespace TrainStation.Framework;
 
 internal class ModConfig
 {
-    public int TicketStationX = 32;
-    public int TicketStationY = 40;
-    public int RailroadWarpX = 32;
-    public int RailroadWarpY = 42;
+    public int TicketStationX { get; set; } = 32;
+    public int TicketStationY { get; set; } = 40;
+    public int RailroadWarpX { get; set; } = 32;
+    public int RailroadWarpY { get; set; } = 42;
 }